requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
pyyaml>=5.4.0
pytest>=7.0.0
